requirements.txt
# PDF Processing
pdfplumber==0.11.7
PyPDF2==3.0.1

# OCR Libraries
pytesseract==0.3.10
Pillow>=10.0.0  # Already installed (11.3.0)
pdf2image==1.17.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Utilities
tqdm==4.66.1
